changed Veinminer to mine nearest blocks first only up to 150 blocks => reduce lag.
