+++ .github/workflows/build.yml
@@ -8,8 +8,8 @@ on: [ pull_request, workflow_dispatch ]
 env: #Environment variables that can later be referenced using ${{ env.MINECRAFT_VERSION }}. These are useful for repeating information and allow for quick changes for new mod updates
   MINECRAFT_VERSION: 1.20.4
   JAVA_VERSION: 17
-  VERSION: 0.0.1
-  RELEASE_NAME: Smacksmisc-1.20.4-0.0.1
+  VERSION: 0.0.2
+  RELEASE_NAME: SmackUtil-1.20.4-0.0.2
 
 permissions:
   contents: write
@@ -48,7 +48,7 @@ jobs:
           name: Artifacts
           path: build/libs/
       - name: publish (CurseForge)
-        if: ${{ runner.os == 'Linux' && matrix.java == '21' }} # Only upload artifacts built from latest java on one OS
+        if: ${{ runner.os == 'Linux' && matrix.java == '21' && github.event_name == 'workflow_dispatch'}} # Only upload artifacts built from latest java on one OS
         uses: Kir-Antipov/mc-publish@v3.3
         with:
           curseforge-id: 952860                                             #The id of your CurseForge project
