+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.2] - 2023.12.22
+
+### Added
+
+#### Veinminer
+
+Drops and item damage are now properly set. Added Tags for different Block groups
+and exact match mode.
+
+### Changed
+
+#### Veinminer
+
+Max 150 block outlines are rendered to reduce lag. Radius is now mode spesific.
+
+### Fixed
+
+Only Push to CurseFroge on manuel Workflow run.
+
 ## [0.0.1] - 2023.12.22
 
 ### Added
