+++ gradle.properties
@@ -7,7 +7,7 @@ minecraft_version=1.20.4
 yarn_mappings=1.20.4+build.3
 loader_version=0.15.3
 # Mod Properties
-mod_version=1.20.4-0.0.1
+mod_version=1.20.4-0.0.2
 maven_group=net.smackplays.smacksutil
 archives_base_name=smacksutil
 # Dependencies
