+++ src/main/java/net/smackplays/smacksutil/mixin/fastplace/MinecraftClientMixin.java
@@ -1,4 +1,4 @@
-package net.smackplays.smacksutil.mixin.FastPlace;
+package net.smackplays.smacksutil.mixin.fastplace;
 
 
 import net.minecraft.client.MinecraftClient;
