+++ src/main/java/net/smackplays/smacksutil/mixin/veinminer/MouseScroll.java
@@ -1,7 +1,7 @@
-package net.smackplays.smacksutil.mixin.VeinMiner;
+package net.smackplays.smacksutil.mixin.veinminer;
 
 import net.minecraft.client.Mouse;
-import net.smackplays.smacksutil.VeinMiner.Miner;
+import net.smackplays.smacksutil.veinminer.Miner;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
