+++ src/main/java/net/smackplays/smacksutil/veinminer/modes/BlockPosComparator.java
@@ -1,20 +1,21 @@
-package net.smackplays.smacksutil.VeinMiner.Modes;
+package net.smackplays.smacksutil.veinminer.modes;
 
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Vec3d;
 
 import java.util.Comparator;
 
 public class BlockPosComparator implements Comparator<BlockPos> {
 
     PlayerEntity player;
-    public BlockPosComparator(PlayerEntity playerIn){
+
+    public BlockPosComparator(PlayerEntity playerIn) {
         player = playerIn;
     }
+
     @Override
     public int compare(BlockPos o1, BlockPos o2) {
-        if(player == null) return 0;
+        if (player == null) return 0;
         double o1ToPlayer = o1.getSquaredDistance(player.getPos());
         double o2ToPlayer = o2.getSquaredDistance(player.getPos());
         if (o1ToPlayer < o2ToPlayer) return -1;
