+++ src/main/java/net/smackplays/smacksutil/veinminer/modes/Mineshaft.java
@@ -1,18 +1,19 @@
-package net.smackplays.smacksutil.VeinMiner.Modes;
+package net.smackplays.smacksutil.veinminer.modes;
 
 import net.minecraft.block.Block;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.registry.tag.TagKey;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Direction;
 import net.minecraft.world.World;
 import net.smackplays.smacksutil.util.ModTags;
 
 import java.util.ArrayList;
 
 @SuppressWarnings("unchecked")
-public class Trees extends VeinMode {
-    public Trees() {
-        ModeName = "Ores";
+public class Mineshaft extends VeinMode {
+    public Mineshaft() {
+        ModeName = "Mineshaft";
     }
 
     @Override
@@ -29,10 +30,15 @@ public class Trees extends VeinMode {
         }
 
         TagKey<Block> tag = null;
-        if (world.getBlockState(sourcePos).isIn(ModTags.Blocks.TREE_BLOCKS)) {
-            tag = ModTags.Blocks.TREE_BLOCKS;
+        if (world.getBlockState(sourcePos).isIn(ModTags.Blocks.VEIN_MINING)) {
+            tag = ModTags.Blocks.VEIN_MINING;
+        } else if (world.getBlockState(sourcePos).isIn(ModTags.Blocks.STONE_BLOCKS)) {
+            tag = ModTags.Blocks.STONE_BLOCKS;
+        } else if (world.getBlockState(sourcePos).isIn(ModTags.Blocks.DIRT_BLOCKS)) {
+            tag = ModTags.Blocks.DIRT_BLOCKS;
         }
-        trees(sourcePos, world, player, isExactMatch, toMatch, tag);
+
+        mineshaft(sourcePos, player.getHorizontalFacing(), radius, player, world, isExactMatch, toMatch, tag);
 
         toBreak.sort(new BlockPosComparator(player));
 
@@ -44,22 +50,18 @@ public class Trees extends VeinMode {
         return (ArrayList<BlockPos>) toBreak.clone();
     }
 
-    private void trees(BlockPos curr, World world, PlayerEntity player, boolean isExactMatch, Block toMatch, TagKey<Block> tag) {
-        if (checked.contains(curr)) {
-            return;
-        } else {
-            checked.add(curr);
-        }
-        if (checkMatch(isExactMatch, curr, world, player, toMatch, tag)) {
-            if (!toBreak.contains(curr)) {
+    public void mineshaft(BlockPos curr, Direction direction, int radius, PlayerEntity player, World world, boolean isExactMatch, Block toMatch, TagKey<Block> tag) {
+        for (int i = 0; i < radius * 2; i++) {
+            if (checkMatch(isExactMatch, curr, world, player, toMatch, tag)) {
                 toBreak.add(curr);
+                for (int j = 1; j < 4; j++) {
+                    if (checkMatch(isExactMatch, curr.offset(direction, j), world, player, toMatch, tag)) {
+                        toBreak.add(curr.offset(direction, j));
+                    }
+                }
+                curr = curr.offset(direction, 1);
             }
-        }
-        BlockPos[] surrounding = getSurrounding(curr);
-        for (BlockPos pos : surrounding) {
-            if (checkMatch(isExactMatch, pos, world, player, toMatch, tag)) {
-                trees(pos, world, player, isExactMatch, toMatch, tag);
-            }
+            curr = curr.down();
         }
     }
 }
