+++ src/main/java/net/smackplays/smacksutil/veinminer/modes/Ores.java
@@ -1,19 +1,18 @@
-package net.smackplays.smacksutil.VeinMiner.Modes;
+package net.smackplays.smacksutil.veinminer.modes;
 
 import net.minecraft.block.Block;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.registry.tag.TagKey;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Direction;
 import net.minecraft.world.World;
 import net.smackplays.smacksutil.util.ModTags;
 
 import java.util.ArrayList;
 
 @SuppressWarnings("unchecked")
-public class Mineshaft extends VeinMode {
-    public Mineshaft() {
-        ModeName = "Mineshaft";
+public class Ores extends VeinMode {
+    public Ores() {
+        ModeName = "Ores";
     }
 
     @Override
@@ -30,15 +29,10 @@ public class Mineshaft extends VeinMode {
         }
 
         TagKey<Block> tag = null;
-        if (world.getBlockState(sourcePos).isIn(ModTags.Blocks.VEIN_MINING)) {
-            tag = ModTags.Blocks.VEIN_MINING;
-        } else if (world.getBlockState(sourcePos).isIn(ModTags.Blocks.STONE_BLOCKS)) {
-            tag = ModTags.Blocks.STONE_BLOCKS;
-        } else if (world.getBlockState(sourcePos).isIn(ModTags.Blocks.DIRT_BLOCKS)) {
-            tag = ModTags.Blocks.DIRT_BLOCKS;
+        if (world.getBlockState(sourcePos).isIn(ModTags.Blocks.ORE_BLOCKS)) {
+            tag = ModTags.Blocks.ORE_BLOCKS;
         }
-
-        mineshaft(sourcePos, player.getHorizontalFacing(), radius, player, world, isExactMatch, toMatch, tag);
+        ores(sourcePos, world, player, isExactMatch, toMatch, tag);
 
         toBreak.sort(new BlockPosComparator(player));
 
@@ -50,18 +44,22 @@ public class Mineshaft extends VeinMode {
         return (ArrayList<BlockPos>) toBreak.clone();
     }
 
-    public void mineshaft(BlockPos curr, Direction direction, int radius, PlayerEntity player, World world, boolean isExactMatch, Block toMatch, TagKey<Block> tag) {
-        for (int i = 0; i < radius * 2; i++) {
-            if (checkMatch(isExactMatch, curr, world, player, toMatch, tag)) {
+    private void ores(BlockPos curr, World world, PlayerEntity player, boolean isExactMatch, Block toMatch, TagKey<Block> tag) {
+        if (checked.contains(curr)) {
+            return;
+        } else {
+            checked.add(curr);
+        }
+        if (checkMatch(isExactMatch, curr, world, player, toMatch, tag)) {
+            if (!toBreak.contains(curr)) {
                 toBreak.add(curr);
-                for (int j = 1; j < 4; j++) {
-                    if (checkMatch(isExactMatch, curr.offset(direction, j), world, player, toMatch, tag)) {
-                        toBreak.add(curr.offset(direction, j));
-                    }
-                }
-                curr = curr.offset(direction, 1);
             }
-            curr = curr.down();
+        }
+        BlockPos[] surrounding = getSurrounding(curr);
+        for (BlockPos pos : surrounding) {
+            if (checkMatch(isExactMatch, pos, world, player, toMatch, tag)) {
+                ores(pos, world, player, isExactMatch, toMatch, tag);
+            }
         }
     }
 }
