+++ src/main/java/net/smackplays/smacksutil/veinminer/modes/ShapelessVertical.java
@@ -1,4 +1,5 @@
-package net.smackplays.smacksutil.VeinMiner.Modes;
+package net.smackplays.smacksutil.veinminer.modes;
+
 
 import net.minecraft.block.Block;
 import net.minecraft.entity.player.PlayerEntity;
@@ -10,15 +11,16 @@ import net.smackplays.smacksutil.util.ModTags;
 import java.util.ArrayList;
 
 @SuppressWarnings("unchecked")
-public class Shapeless extends VeinMode {
-    public Shapeless() {
-        ModeName = "Shapeless";
-        MAX_RADIUS = 4;
+public class ShapelessVertical extends VeinMode {
+    public ShapelessVertical() {
+        ModeName = "ShapelessVertical";
+        MAX_RADIUS = 5;
     }
 
     @Override
     public ArrayList<BlockPos> getBlocks(World world, PlayerEntity player, BlockPos sourcePos, int radius, boolean isExactMatch) {
         if (world == null || player == null || sourcePos == null) return (ArrayList<BlockPos>) toBreak.clone();
+        oldToBreak = (ArrayList<BlockPos>) toBreak.clone();
         toBreak.clear();
         checked.clear();
         Block toMatch = world.getBlockState(sourcePos).getBlock();
@@ -35,7 +37,7 @@ public class Shapeless extends VeinMode {
             tag = ModTags.Blocks.DIRT_BLOCKS;
         }
 
-        shapeless(sourcePos, sourcePos, radius, world, player, isExactMatch, toMatch, tag);
+        shapeless_vert(sourcePos, sourcePos, radius, player, world, isExactMatch, toMatch, tag);
 
         toBreak.sort(new BlockPosComparator(player));
 
@@ -47,14 +49,14 @@ public class Shapeless extends VeinMode {
         return (ArrayList<BlockPos>) toBreak.clone();
     }
 
-    private void shapeless(BlockPos curr, BlockPos sourcePos, int radius, World world,
-                           PlayerEntity player, boolean isExactMatch, Block toMatch, TagKey<Block> tag) {
+    private void shapeless_vert(BlockPos curr, BlockPos sourcePos, int radius, PlayerEntity player,
+                                World world, boolean isExactMatch, Block toMatch, TagKey<Block> tag) {
         if (curr.getX() > sourcePos.getX() + radius
                 || curr.getX() < sourcePos.getX() - radius) {
             return;
         }
         if (curr.getY() > sourcePos.getY() + radius
-                || curr.getY() < sourcePos.getY() - radius) {
+                || curr.getY() < player.getY()) {
             return;
         }
         if (curr.getZ() > sourcePos.getZ() + radius
@@ -70,9 +72,10 @@ public class Shapeless extends VeinMode {
             toBreak.add(curr);
         }
         BlockPos[] surrounding = getSurrounding(curr);
+
         for (BlockPos pos : surrounding) {
             if (checkMatch(isExactMatch, pos, world, player, toMatch, tag)) {
-                shapeless(pos, sourcePos, radius, world, player, isExactMatch, toMatch, tag);
+                shapeless_vert(pos, sourcePos, radius, player, world, isExactMatch, toMatch, tag);
             }
         }
     }
